service den

+++ SpringBootJPA_ISCmanagement_project/src/main/java/com/ISC/project/dao/CompanyRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.ISC.project.model.Company;
 
-public interface CompanyRepository extends JpaRepository<Company, Integer>{
+public interface CompanyRepository extends JpaRepository<Company, Long>{
 
 }
