+++ SpringBootJPA_ISCmanagement_project/src/main/java/com/ISC/project/dao/LecturerRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.ISC.project.model.Lecturer;
 
-public interface LecturerRepository extends JpaRepository<Lecturer, Integer>{
+public interface LecturerRepository extends JpaRepository<Lecturer, Long>{
 
 }
