+++ SpringBootJPA_ISCmanagement_project/src/main/java/com/ISC/project/dao/MajorRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.ISC.project.model.Major;
 
-public interface MajorRepository extends JpaRepository<Major, Integer>{
+public interface MajorRepository extends JpaRepository<Major, Long>{
 
 }
