+++ SpringBootJPA_ISCmanagement_project/src/main/java/com/ISC/project/dao/SubjectRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.ISC.project.model.Subject;
 
-public interface SubjectRepository extends JpaRepository<Subject, Integer>{
+public interface SubjectRepository extends JpaRepository<Subject, Long>{
 
 }
